src/test/java/com/humaorie/strtotime/strtotimeTest.java
package com.humaorie.strtotime;

/*
 * strtotime, (c) 2009 Davide Angelocola <davide.angelocola@gmail.com>
 * http://bitbucket.org/dfa/strtotime
 *
 * This program is free software: you can redistribute it and/or modify
 * it under the terms of the GNU Lesser Public License as published by
 * the Free Software Foundation, either version 3 of the License, or
 * (at your option) any later version.
 *
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU Lesser Public License for more details.
 *
 * You should have received a copy of the GNU General Public License
 * along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
import static com.humaorie.strtotime.strtotime.*;
import java.util.Date;
import org.joda.time.DateTime;
import org.junit.Test;
import static org.junit.Assert.*;
import static org.hamcrest.CoreMatchers.*;

public class strtotimeTest {

    @Test
    public void now() {
        // this test always fail
        //assertThat(strtotime("now"), is(new DateTime().toDate()));
        assertThat(strtotime("now"), notNullValue());
    }
    @Test

    public void nowWithBlanks() {
        // this test always fail
        //assertThat(strtotime("now"), is(new DateTime().toDate()));
        assertThat(strtotime("  now   "), notNullValue());
    }

    @Test
    public void today() {
        // this test always fail
        //assertThat(strtotime("today"), is(new DateTime().toDate()));
        assertThat(strtotime("today"), notNullValue());
    }

    @Test
    public void todayWithBlanks() {
        // this test always fail
        //assertThat(strtotime("today "), is(new DateTime().toDate()));
        assertThat(strtotime("  today  "), notNullValue());
    }

    @Test
    public void tomorrow() {
        assertThat(strtotime("tomorrow"), is(new DateTime().plusDays(1).toDate()));
    }

    @Test
    public void yesterday() {
        assertThat(strtotime("yesterday"), is(new DateTime().minusDays(1).toDate()));
    }

    @Test
    public void days() {
        assertThat(strtotime("3 days"), is(new DateTime().plusDays(3).toDate()));
    }

    @Test
    public void weeks() {
        assertThat(strtotime("3 weeks"), is(new DateTime().plusWeeks(3).toDate()));
    }

    @Test
    public void customizeMatchers() {
        strtotime.registerMatcher(new Matcher() {

            @Override
            public Date tryConvert(String input) {
                if ("my format".equals(input)) {
                    return new Date(0);
                } else {
                    return null;
                }
            }
        });

        assertThat(strtotime("my format"), notNullValue());
        assertThat(strtotime("my format"), is(new Date(0)));
    }

    @Test
    public void customizeOverrides() {
        strtotime.registerMatcher(new strtotime.Matcher() { // overrides "now"

            @Override
            public Date tryConvert(String input) {
                if ("now".equals(input)) {
                    return new Date(0);
                } else {
                    return null;
                }
            }
        });

        assertThat(strtotime("now"), is(new Date(0)));
    }

    @Test
    public void yearMonthDay() {
        assertThat(strtotime("2009 12 25"), is(new DateTime(2009, 12, 25, 0, 0, 0, 0).toDate()));
    }

    @Test
    public void error() {
        assertThat(strtotime("unsupported date format"), nullValue());
    }
}
